Raise an error if the user try to import ASA 8
Actually we have too much trouble with recent OS
and ASA 8 with a success rate getting worse each day.
In 1.4.1 we will promote the switch to ASAv sadly it's not
possible to convert from ASA 8 to ASAv.

This PR raise an error making clear to the user that this will
not work.